functions/load-all/requirements.txt
functions-framework
google-cloud-storage
google-cloud-secret-manager
duckdb
pandas
python-dateutil